<feat>: 로봇 버튼 구현
